+++ src/importexport/musicxml/internal/musicxml/exportxml.cpp
@@ -124,7 +124,6 @@
 #include "engraving/dom/volta.h"
 #include "engraving/dom/whammybar.h"
 
-#include "musicxml.h"
 #include "musicxmlfonthandler.h"
 #include "musicxmlsupport.h"
 #include "musicxmltypes.h"
@@ -157,6 +156,8 @@ namespace mu::engraving {
 #else
 #define clefDebug(...) {}
 #endif
+
+constexpr int MAX_PART_GROUPS  = 8;
 
 //---------------------------------------------------------
 //   typedefs
