+++ src/importexport/musicxml/internal/musicxml/importmxmlnoteduration.h
@@ -26,7 +26,6 @@
 #include "engraving/dom/durationtype.h"
 #include "engraving/types/fraction.h"
 #include "importmxmlpass1.h"
-#include "musicxmltypes.h"
 
 namespace mu::engraving {
 class MxmlLogger;
