+++ src/importexport/musicxml/internal/musicxml/importmxmlpass1.h
@@ -36,6 +36,7 @@
 
 namespace mu::engraving {
 class Score;
+class VoiceOverlapDetector;
 
 //---------------------------------------------------------
 //   PageFormat
@@ -55,53 +56,6 @@ struct PageFormat {
 
 typedef std::pair<int, int> StartStop;
 typedef std::vector<StartStop> StartStopList;
-
-//---------------------------------------------------------
-//   NoteList
-//---------------------------------------------------------
-
-/**
- List of note start/stop times in a voice in all staves.
-*/
-
-class NoteList
-{
-public:
-    NoteList();
-    void addNote(const int startTick, const int endTick, const size_t staff);
-    void dump(const int& voice) const;
-    bool stavesOverlap(const int staff1, const int staff2) const;
-    bool anyStaffOverlaps() const;
-private:
-    std::vector<StartStopList> _staffNoteLists;   // The note start/stop times in all staves
-    bool notesOverlap(const StartStop& n1, const StartStop& n2) const;
-};
-
-//---------------------------------------------------------
-//   VoiceOverlapDetector
-//---------------------------------------------------------
-
-/**
- Detect overlap in a voice, which is when a voice has two or more notes
- active at the same time. In theory this should not happen, as voices
- only move forward in time, but Sibelius 7 reuses voice numbers in multi-
- staff parts, which leads to overlap.
-
- Current implementation does not detect voice overlap within a staff,
- but only between staves.
-*/
-
-class VoiceOverlapDetector
-{
-public:
-    VoiceOverlapDetector();
-    void addNote(const int startTick, const int endTick, const int& voice, const int staff);
-    void dump() const;
-    void newMeasure();
-    bool stavesOverlap(const int& voice) const;
-private:
-    std::map<int, NoteList> _noteLists;   // The notelists for all the voices
-};
 
 //---------------------------------------------------------
 //   MxmlOctaveShiftDesc
