--- src/importexport/musicxml/internal/musicxml/musicxml.h
@@ -1,31 +1,0 @@
-/*
- * SPDX-License-Identifier: GPL-3.0-only
- * MuseScore-Studio-CLA-applies
- *
- * MuseScore Studio
- * Music Composition & Notation
- *
- * Copyright (C) 2021 MuseScore Limited
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-#ifndef __MUSICXML_H__
-#define __MUSICXML_H__
-
-namespace mu::engraving {
-const int MAX_LYRICS       = 16;
-const int MAX_PART_GROUPS  = 8;
-const int MAX_NUMBER_LEVEL = 16; // maximum number of overlapping MusicXML objects
-} // namespace Ms
-#endif
